# ZClub/docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: ZClub 积分系统 API
  description: ZClub 积分系统的 RESTful API 文档
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: 本地开发环境
paths:
...
  /auth/register:
    post:
      summary: 用户注册
      tags:
        - 认证
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: 
                  type: string
                email: 
                  type: string
                password: 
                  type: string
                role: 
                  type: string
                  enum: [admin, parent, child]
                parent_id: 
                  type: string
                family_id: 
                  type: string
                age: 
                  type: integer
      responses:
        '201':
          description: 注册成功
        '400':
          description: 注册失败
  /auth/login:
    post:
      summary: 用户登录
      tags:
        - 认证
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: 
                  type: string
                password: 
                  type: string
      responses:
        '200':
          description: 登录成功
        '401':
          description: 登录失败

  # 家庭相关 API
...
  /tasks:
    get:
      summary: 获取任务列表
      tags:
        - 任务
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 成功

    post:
      summary: 创建任务
      tags:
        - 任务
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: 
                  type: string
                description: 
                  type: string
                points: 
                  type: integer
                assignee_id: 
                  type: string
                due_date: 
                  type: string
                  format: date-time
                recurrence: 
                  type: string
      responses:
        '201':
          description: 创建成功
...
  /tasks/{id}/complete:
    post:
      summary: 完成任务
      tags:
        - 任务
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 提交成功
...
  /rewards:
    get:
      summary: 获取奖励列表
      tags:
        - 奖励
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 成功

    post:
      summary: 创建奖励
      tags:
        - 奖励
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: 
                  type: string
                description: 
                  type: string
                points_required: 
                  type: integer
                stock: 
                  type: integer
      responses:
        '201':
          description: 创建成功
...
  /rewards/{id}/redeem:
    post:
      summary: 兑换奖励
      tags:
        - 奖励
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 申请成功
...
  /notifications:
    get:
      summary: 获取通知列表
      tags:
        - 通知
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 成功
  /notifications/{id}/read:
    put:
      summary: 标记通知为已读
      tags:
        - 通知
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功

  # 报表相关 API
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT